TINY-9532: move Tree demo code from code demo to silver demo
