+++ modules/tinymce/src/themes/silver/demo/html/demo/tree-demo.html
@@ -1,0 +1,21 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+  <meta name="viewport" content="width=device-width, initial-scale=1.0, user-scalable=no, maximum-scale=1.0">
+  <title>silver-theme Mock Demo Page</title>
+  <script src="../../../../../../js/tinymce/tinymce.js"></script>
+  <script src="../../../../../../scratch/demos/themes/silver/demo.js"></script>
+
+</head>
+
+<body>
+  <h2>Mock Tinymce Modern with silver-theme Demo Page</h2>
+  <textarea class='tinymce'></textarea>
+
+  <script>
+    demos.TreeDemo();
+  </script>
+</body>
+
+</html>
